Add reset filter and clear filter when different player is fetched.
